requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.1.7
django-ckeditor==6.5.1
django-dotenv==1.4.2
django-js-asset==2.0.0
django-restframework==0.0.1
djangorestframework==3.14.0
Flask==2.3.2
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
mysqlclient==2.1.1
Pillow==9.5.0
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.30.0
sqlparse==0.4.3
twilio==8.2.0
tzdata==2022.7
urllib3==2.0.2
Werkzeug==2.3.3
yarl==1.9.2
